backend/requirements.txt
djangorestframework
pymysql
django-cors-headers
bcrypt
pyjwt
